+++ ValleyBikeSimA45/src/FoundingMember.java
@@ -20,7 +20,7 @@ public class FoundingMember extends Membership {
 	@Override
 	public float getChargeForRide(int rideDurationInMinutes) {
 		
-		int totalAmount = 0;
+		float totalAmount = 0;
 		
 		// if the user's ride lasted 60 mins or more
 		if (rideDurationInMinutes >= 60) {
