+++ ValleyBikeSimA45/src/Monthly.java
@@ -20,7 +20,7 @@ public class Monthly extends Membership {
 	@Override
 	public float getChargeForRide(int rideDurationInMinutes) {
 		
-		int totalAmount = 0;
+		float totalAmount = 0;
 		
 		// if the user's ride lasted 45 mins or more
 		if (rideDurationInMinutes >= 45) {
